Feature/setup

+++ reconstruct_piano/m4a2wav/convert.py
@@ -4,25 +4,26 @@ import argparse
 from pydub import AudioSegment
 
 BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-CONVERT_PATH = os.path.join(BASE_DIR, 'm4a2wav')
+PIANO_PATH = os.path.join(BASE_DIR, 'detect_frequency')
+CONVERT_PATH = os.path.join(PIANO_PATH, 'audiofile')
 
 
 def m4a2wave(music_name):
     formats_to_convert = ['.m4a']
-
     for (dirpath, dirnames, filenames) in os.walk(CONVERT_PATH):
         for filename in filenames:
-            if filename.endswith(tuple(formats_to_convert)):
+            if filename == music_name:
+                if filename.endswith(tuple(formats_to_convert)):
 
-                filepath = dirpath + '/' + filename
-                (path, file_extension) = os.path.splitext(filepath)
-                file_extension_final = file_extension.replace('.', '')
-                try:
-                    track = AudioSegment.from_file(filepath,
+                    filepath = dirpath + '/' + filename
+                    (path, file_extension) = os.path.splitext(filepath)
+                    file_extension_final = file_extension.replace('.', '')
+                    try:
+                        track = AudioSegment.from_file(filepath,
                                                    file_extension_final)
-                    wav_filename = filename.replace(file_extension_final, 'wav')
-                    wav_path = dirpath + '/' + wav_filename
-                    file_handle = track.export(wav_path, format='wav')
-                    os.remove(filepath)
-                except:
-                    print("ERROR CONVERTING " + str(filepath))
+                        wav_filename = filename.replace(file_extension_final, 'wav')
+                        wav_path = dirpath + '/' + wav_filename
+                        file_handle = track.export(wav_path, format='wav')
+                        os.remove(filepath)
+                    except:
+                        print("ERROR CONVERTING " + str(filepath))
